refactor: 요금 계산 로직 리팩토링
